Removing duplicated code.
