+++ src/main/java/ai/philterd/phileas/policy/filters/AbstractDictionaryBackedFilter.java
@@ -1,0 +1,82 @@
+/*
+ *     Copyright 2025 Philterd, LLC @ https://www.philterd.ai
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *          http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ai.philterd.phileas.policy.filters;
+
+import ai.philterd.phileas.model.filtering.SensitivityLevel;
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+import java.util.List;
+
+/**
+ * Options shared by the filters that match against a word list: first name, surname, city,
+ * county, state, and hospital. Terms given here replace the list bundled in the jar.
+ */
+public abstract class AbstractDictionaryBackedFilter extends AbstractFilter {
+
+    @SerializedName("terms")
+    @Expose
+    private List<String> terms;
+
+    @SerializedName("fuzzy")
+    @Expose
+    private boolean fuzzy = false;
+
+    @SerializedName("sensitivity")
+    @Expose
+    private String sensitivity = SensitivityLevel.MEDIUM.getName();
+
+    @SerializedName("capitalized")
+    @Expose
+    private boolean capitalized = false;
+
+    public List<String> getTerms() {
+        return terms;
+    }
+
+    public void setTerms(List<String> terms) {
+        this.terms = terms;
+    }
+
+    public SensitivityLevel getSensitivityLevel() {
+        return SensitivityLevel.fromName(sensitivity);
+    }
+
+    public String getSensitivity() {
+        return sensitivity;
+    }
+
+    public void setSensitivity(String sensitivity) {
+        this.sensitivity = sensitivity;
+    }
+
+    public boolean isCapitalized() {
+        return capitalized;
+    }
+
+    public void setCapitalized(boolean capitalized) {
+        this.capitalized = capitalized;
+    }
+
+    public boolean isFuzzy() {
+        return fuzzy;
+    }
+
+    public void setFuzzy(boolean fuzzy) {
+        this.fuzzy = fuzzy;
+    }
+
+}
