+++ src/main/java/ai/philterd/phileas/policy/filters/County.java
@@ -15,34 +15,17 @@
  */
 package ai.philterd.phileas.policy.filters;
 
-import ai.philterd.phileas.model.filtering.SensitivityLevel;
 import ai.philterd.phileas.services.strategies.dynamic.CountyFilterStrategy;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 import java.util.List;
 
-public class County extends AbstractFilter {
+public class County extends AbstractDictionaryBackedFilter {
 
     @SerializedName("countyFilterStrategies")
     @Expose
     private List<CountyFilterStrategy> countyFilterStrategies;
-
-    @SerializedName("terms")
-    @Expose
-    private List<String> terms;
-
-    @SerializedName("fuzzy")
-    @Expose
-    private boolean fuzzy = false;
-
-    @SerializedName("sensitivity")
-    @Expose
-    private String sensitivity = SensitivityLevel.MEDIUM.getName();
-
-    @SerializedName("capitalized")
-    @Expose
-    private boolean capitalized = false;
 
     public List<CountyFilterStrategy> getCountyFilterStrategies() {
         return countyFilterStrategies;
@@ -50,42 +33,6 @@ public class County extends AbstractFilter {
 
     public void setCountyFilterStrategies(List<CountyFilterStrategy> countyFilterStrategies) {
         this.countyFilterStrategies = countyFilterStrategies;
-    }
-
-    public SensitivityLevel getSensitivityLevel() {
-        return SensitivityLevel.fromName(sensitivity);
-    }
-
-    public String getSensitivity() {
-        return sensitivity;
-    }
-
-    public void setSensitivity(String sensitivity) {
-        this.sensitivity = sensitivity;
-    }
-
-    public boolean isCapitalized() {
-        return capitalized;
-    }
-
-    public void setCapitalized(boolean capitalized) {
-        this.capitalized = capitalized;
-    }
-
-    public boolean isFuzzy() {
-        return fuzzy;
-    }
-
-    public void setFuzzy(boolean fuzzy) {
-        this.fuzzy = fuzzy;
-    }
-
-    public List<String> getTerms() {
-        return terms;
-    }
-
-    public void setTerms(List<String> terms) {
-        this.terms = terms;
     }
 
 }
