+++ src/main/java/ai/philterd/phileas/policy/filters/State.java
@@ -15,34 +15,17 @@
  */
 package ai.philterd.phileas.policy.filters;
 
-import ai.philterd.phileas.model.filtering.SensitivityLevel;
 import ai.philterd.phileas.services.strategies.dynamic.StateFilterStrategy;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 import java.util.List;
 
-public class State extends AbstractFilter {
+public class State extends AbstractDictionaryBackedFilter {
 
     @SerializedName("stateFilterStrategies")
     @Expose
     private List<StateFilterStrategy> stateFilterStrategies;
-
-    @SerializedName("terms")
-    @Expose
-    private List<String> terms;
-
-    @SerializedName("fuzzy")
-    @Expose
-    private boolean fuzzy = false;
-
-    @SerializedName("sensitivity")
-    @Expose
-    private String sensitivity = SensitivityLevel.MEDIUM.getName();
-
-    @SerializedName("capitalized")
-    @Expose
-    private boolean capitalized = false;
 
     public List<StateFilterStrategy> getStateFilterStrategies() {
         return stateFilterStrategies;
@@ -50,42 +33,6 @@ public class State extends AbstractFilter {
 
     public void setStateFilterStrategies(List<StateFilterStrategy> stateFilterStrategies) {
         this.stateFilterStrategies = stateFilterStrategies;
-    }
-
-    public SensitivityLevel getSensitivityLevel() {
-        return SensitivityLevel.fromName(sensitivity);
-    }
-
-    public String getSensitivity() {
-        return sensitivity;
-    }
-
-    public void setSensitivity(String sensitivity) {
-        this.sensitivity = sensitivity;
-    }
-
-    public boolean isCapitalized() {
-        return capitalized;
-    }
-
-    public void setCapitalized(boolean capitalized) {
-        this.capitalized = capitalized;
-    }
-
-    public boolean isFuzzy() {
-        return fuzzy;
-    }
-
-    public void setFuzzy(boolean fuzzy) {
-        this.fuzzy = fuzzy;
-    }
-
-    public List<String> getTerms() {
-        return terms;
-    }
-
-    public void setTerms(List<String> terms) {
-        this.terms = terms;
     }
 
 }
