+++ src/main/java/ai/philterd/phileas/policy/filters/Surname.java
@@ -15,34 +15,17 @@
  */
 package ai.philterd.phileas.policy.filters;
 
-import ai.philterd.phileas.model.filtering.SensitivityLevel;
 import ai.philterd.phileas.services.strategies.dynamic.SurnameFilterStrategy;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 import java.util.List;
 
-public class Surname extends AbstractFilter {
+public class Surname extends AbstractDictionaryBackedFilter {
 
     @SerializedName("surnameFilterStrategies")
     @Expose
     private List<SurnameFilterStrategy> surnameFilterStrategies;
-
-    @SerializedName("terms")
-    @Expose
-    private List<String> terms;
-
-    @SerializedName("fuzzy")
-    @Expose
-    private boolean fuzzy = false;
-
-    @SerializedName("sensitivity")
-    @Expose
-    private String sensitivity = SensitivityLevel.MEDIUM.getName();
-
-    @SerializedName("capitalized")
-    @Expose
-    private boolean capitalized = false;
 
     public List<SurnameFilterStrategy> getSurnameFilterStrategies() {
         return surnameFilterStrategies;
@@ -50,42 +33,6 @@ public class Surname extends AbstractFilter {
 
     public void setSurnameFilterStrategies(List<SurnameFilterStrategy> surnameFilterStrategies) {
         this.surnameFilterStrategies = surnameFilterStrategies;
-    }
-
-    public SensitivityLevel getSensitivityLevel() {
-        return SensitivityLevel.fromName(sensitivity);
-    }
-
-    public String getSensitivity() {
-        return sensitivity;
-    }
-
-    public void setSensitivity(String sensitivity) {
-        this.sensitivity = sensitivity;
-    }
-
-    public boolean isCapitalized() {
-        return capitalized;
-    }
-
-    public void setCapitalized(boolean capitalized) {
-        this.capitalized = capitalized;
-    }
-
-    public boolean isFuzzy() {
-        return fuzzy;
-    }
-
-    public void setFuzzy(boolean fuzzy) {
-        this.fuzzy = fuzzy;
-    }
-
-    public List<String> getTerms() {
-        return terms;
-    }
-
-    public void setTerms(List<String> terms) {
-        this.terms = terms;
     }
 
 }
